Walk on ground
select circle, edit mode than control points and set handle type to vector. Transform pivot point from median point to 3d cursor (which is at the center now) scale that bottom circle point on Z axis to 0. This will bring the point to the height of X axes. Now we have a half circle and the leg is walking on the ground.