Random movement
Split the screen and open graph editor, select the body and press I  and select Location and rotation (They are added to the graph editor now).

BodyActions dropdown and select X location with N key open menu and select modifiers and add Noise modifier.

Adjust Scale and Phase (I did 15 & 30) Beside add modifier copy F modifier and than select Y Location and past it afterwards adjust the phase. Do the same for Z axis.

Past the same modifier to all rotations with a lower Strength.

Play it and adjust strength for all rotations and adjust phase and strength for locations.