Moving across the screen
Add empty cube and scale it so everything fit inside and name it Character controller.

Select Direction Controller, all paths the body and than Shift select the Character Controller parent it using Ctrl + P and select Object.

Move the Character controller to a location on X axis and insert key frame with I and than move it to a different location in direction of X axis and another key frame. Tricky part is to adjust it by moving one of the keyframes so the character is not sliding instead walking.

Again hover the mouse above timeline and do shift E and Linear extrapolation to move it indefinitely